pyproject.toml
[project]
name = "omnitech-chronicles-bot"
version = "1.0.0"
description = "OmniTech Chronicles — Telegram visual novel bot (Episode 1)"
requires-python = ">=3.11"
dependencies = [
    "pytelegrambotapi>=4.20.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
